I think i finally got this bug sorted out.
I hate when this happens to me, but i am also very pleased for having done it.

It only makes sense to replace all the instances of (CAR *SYMLIS) with *SYMLIS.

Well, now i'll take a break for a bit, and later maybe i'll add commentary and
refacor the code and the variable names.